+++ server/component/style/input/InputView.php
@@ -41,13 +41,7 @@ class InputView extends FormFieldView
      * DB field 'format' ('').
      * Format field for the input
      */
-    private $format;
-
-    /**
-     * DB field 'locked_after_submit' ('').
-     * If selected and if the field is used in a form that is not `is_log`, once the value is set, the field will not be able to be edited anymore.
-     */
-    private $locked_after_submit;
+    private $format;    
 
     /* Constructors ***********************************************************/
 
@@ -75,8 +69,7 @@ class InputView extends FormFieldView
             if($this->label == "") $this->label = "&zwnj;";
         }
         $this->data_config = $this->model->get_db_field("data_config");
-        $this->value = $this->model->get_db_field("value", "");
-        $this->locked_after_submit = $this->model->get_db_field("locked_after_submit", 0);
+        $this->value = $this->model->get_db_field("value", "");        
         if($this->data_config){
             $this->retrieve_data();
         }